feat(dev): expose restart function on ViteDevServer (#5723)
